requirements.txt
numpy==2.2.2
pandas==2.2.3
matplotlib==3.10.0
python-binance==1.0.27
python-dotenv==1.0.1